Refactor CLI commands to root submodule
