--- fastapi_mvc/cli/commands/__init__.py
@@ -1,1 +1,0 @@
-"""fastapi-mvc."""
